move defaultStop
