+++ resources/camino-creator/common/defaultStop.js
